Added numerical integration test to spring_mass_system_test, and enhanced spring_mass_system to report energy. Where I ran into API problems or inconsistencies, I updated the API. Specifics: Renamed Time struct to TimeStep to make it easier to talk about time; added get_time() and set_time() methods to Context. Renamed Dynamics() to GetDerivatives() and changed its return type to ContinuousState. Renamed AllocateStateDerivatives() to AllocateDerivatives() for consistency. Renamed Output() to GetOutput() for consistency.

+++ drake/systems/framework/basic_vector.h
@@ -21,7 +21,7 @@ namespace systems {
 template <typename T>
 class BasicVector : public VectorInterface<T> {
  public:
-  explicit BasicVector(int size)
+  explicit BasicVector(ptrdiff_t size)
       : values_(VectorX<T>::Constant(
             size, std::numeric_limits<
                       typename Eigen::NumTraits<T>::Real>::quiet_NaN())) {}
